Clean up and resolve packr2 issues
This patch addresses issues with the build pipeline caused by an invalid import. Profiling was also added.
